+++ src/app/views/Notifications.svelte
@@ -9,7 +9,7 @@
   import Note from "src/app/shared/Note.svelte"
   import GroupAlert from "src/app/shared/GroupAlert.svelte"
   import GroupRequest from "src/app/shared/GroupRequest.svelte"
-  import OnboardingTasks from "src/app/views/OnboardingTasks.svelte"
+  import OnboardingTasks from "src/app/shared/OnboardingTasks.svelte"
   import NotificationReactions from "src/app/views/NotificationReactions.svelte"
   import NotificationMention from "src/app/views/NotificationMention.svelte"
   import NotificationReplies from "src/app/views/NotificationReplies.svelte"