+++ CHANGELOG.md
@@ -16,7 +16,8 @@
 - [x] Fix group search
 - [x] Join invite relays in background during onboarding
 - [x] Republish user profile data when joining a new relay
-- [x] Create new note publish dialog
+- [x] Create new note publish status dialog
+- [x] Add wallet setup onboarding item
 
 # 0.4.4